requirements.txt

reflex==0.8.2
sqlalchemy
alembic
aiosqlite
pymysql
fastapi
uvicorn
bcrypt
bleach
slowapi
cryptography
python-dotenv